feat: AddTokenReview component has its related functions inside to be more re-useful
