+++ src/frontend/src/eth/components/tokens/AddTokenReview.svelte
@@ -12,9 +12,19 @@
 	import ButtonGroup from '$lib/components/ui/ButtonGroup.svelte';
 	import AddTokenWarning from '$lib/components/tokens/AddTokenWarning.svelte';
 	import { tokenWithFallback } from '$lib/derived/token.derived';
+	import { authStore } from '$lib/stores/auth.store';
+	import { nullishSignOut } from '$lib/services/auth.services';
+	import { ProgressStepsAddToken } from '$lib/enums/progress-steps';
+	import { addUserToken } from '$lib/api/backend.api';
+	import { selectedChainId, selectedEthereumNetwork } from '$eth/derived/network.derived';
+	import { mapErc20Token } from '$eth/utils/erc20.utils';
+	import type { WizardModal } from '@dfinity/gix-components';
 
+	export let modal: WizardModal;
+	export let saveProgressStep: ProgressStepsAddToken;
 	export let contractAddress = '';
-	export let metadata: Erc20Metadata | undefined;
+
+	let metadata: Erc20Metadata | undefined;
 
 	onMount(async () => {
 		if (
@@ -71,6 +81,67 @@
 	$: invalid = isNullishOrEmpty(contractAddress) || isNullish(metadata);
 
 	const dispatch = createEventDispatcher();
+
+	const save = async () => {
+		if (isNullishOrEmpty(contractAddress)) {
+			toastsError({
+				msg: { text: $i18n.tokens.error.invalid_contract_address }
+			});
+			return;
+		}
+
+		if (isNullish(metadata)) {
+			toastsError({
+				msg: { text: $i18n.tokens.error.no_metadata }
+			});
+			return;
+		}
+
+		if (isNullish($authStore.identity)) {
+			await nullishSignOut();
+			return;
+		}
+
+		modal.next();
+
+		try {
+			saveProgressStep = ProgressStepsAddToken.SAVE;
+
+			await addUserToken({
+				identity: $authStore.identity,
+				token: {
+					chain_id: $selectedChainId,
+					contract_address: contractAddress,
+					symbol: [],
+					decimals: [],
+					version: []
+				}
+			});
+
+			saveProgressStep = ProgressStepsAddToken.UPDATE_UI;
+
+			erc20TokensStore.add(
+				mapErc20Token({
+					address: contractAddress,
+					exchange: 'ethereum',
+					category: 'custom',
+					network: $selectedEthereumNetwork,
+					...metadata
+				})
+			);
+
+			saveProgressStep = ProgressStepsAddToken.DONE;
+
+			setTimeout(() => dispatch('icClose'), 750);
+		} catch (err: unknown) {
+			toastsError({
+				msg: { text: $i18n.tokens.error.unexpected },
+				err
+			});
+
+			modal.back();
+		}
+	};
 </script>
 
 <div class="stretch">
@@ -117,7 +188,7 @@
 		class="primary block flex-1"
 		disabled={invalid}
 		class:opacity-10={invalid}
-		on:click={() => dispatch('icSave')}
+		on:click={save}
 	>
 		{$i18n.tokens.import.text.add_the_token}
 	</button>
